splitto multiple files
