+++ tests/__init__.py
@@ -1,0 +1,27 @@
+import testRoles
+import ctypes,unittest
+libcap = ctypes.cdll.LoadLibrary("libcap.so")
+
+libcap.cap_get_proc.argtypes = []
+libcap.cap_get_proc.restype = ctypes.c_void_p
+libcap.cap_to_text.restype = ctypes.c_char_p
+libcap.cap_free.restype = ctypes.c_void_p
+cap_p = libcap.cap_get_proc()
+currentcaps = libcap.cap_to_text(cap_p, None)
+
+test_Roles = (testRoles.TestFindUserRoles, testRoles.TestFindGroupRoles, testRoles.TestFindGroupNoRole) 
+
+def load_tests(loader, tests, pattern):
+    suite = unittest.TestSuite()
+    for test_class in test_Roles:
+        tests = loader.loadTestsFromTestCase(test_class)
+        suite.addTests(tests)
+    return suite
+
+if __name__ == '__main__':
+    if str(currentcaps).find("cap_dac_override") < 0:
+        print("please run this with sr")
+        exit(-1)
+    unittest.main()
+
+libcap.cap_free(currentcaps)
